+++ src/test/resources/evaltest.xqy
@@ -1,0 +1,3 @@
+xquery version "3.0";
+
+<doc>test</doc>
